runtime: fix the Linux build — nova_mono_ms return type + missing <sys/stat.h>
Two errors that only fire on Linux, so the Windows syscall-shim work did not
surface them:

- concurrency.cpp declared `extern "C" long long nova_mono_ms(void)` while
  core.cpp defines it as int64_t. Those are the same type on macOS/Windows but
  NOT on Linux x86_64, where int64_t is `long` — C++ then rejects the pair as
  "functions that differ only in their return type cannot be overloaded".
- io.cpp calls ::mkdir (iso_setup_rootfs) and relied on <sys/stat.h> arriving
  transitively, which happens on macOS but not on Linux.

+++ src/runtime/concurrency.cpp
@@ -11,6 +11,7 @@
 #include <cstdio>
 #include <chrono>
 #include <condition_variable>
+#include <cstdint>
 #include <cstdlib>
 #include <memory>
 #include <mutex>
@@ -27,7 +28,10 @@ extern "C" long long __nova_main(void);
 extern "C" long long nova_reactor_current(void);
 extern "C" void nova_reactor_set_timer(long long handle, long long ms);
 extern "C" void nova_reactor_cancel_timer(long long handle);
-extern "C" long long nova_mono_ms(void);
+// int64_t, NOT `long long`: core.cpp defines this as int64_t, which is `long long` on macOS but
+// `long` on Linux x86_64 — declaring it `long long` here makes the two differ only in return type,
+// which C++ rejects outright ("functions that differ only in their return type cannot be overloaded").
+extern "C" int64_t nova_mono_ms(void);
 
 extern "C" {
